Correct symmetry factors:
The symmetry factors stemming from using a non-redundant basis (such as in WCxf
vs. using a redundant basis where the Wilson coefficients have the same flavour
index symmetries as the operators had not yet been taken into account in the
flavour rotations required in the matching from SMEFT to WET as well as in the
translations from the Warsaw-up to the Warsaw basis.

This is fixed now.